Vector and Color editors

+++ Core/Injection/Memory/VectorMemory.cs
@@ -23,9 +23,9 @@ namespace ConceptMatrix.Injection.Memory
 
 		protected override void Write(Vector value, ref byte[] data)
 		{
-			Array.Copy(BitConverter.GetBytes((float)value.X), data, 4);
-			Array.Copy(BitConverter.GetBytes((float)value.Y), 0, data, 4, 4);
-			Array.Copy(BitConverter.GetBytes((float)value.Z), 0, data, 8, 4);
+			Array.Copy(BitConverter.GetBytes(value.X), data, 4);
+			Array.Copy(BitConverter.GetBytes(value.Y), 0, data, 4, 4);
+			Array.Copy(BitConverter.GetBytes(value.Z), 0, data, 8, 4);
 		}
 	}
 }
